feat: new email system
