+++ src/services/email/generalMails.ts
@@ -1,0 +1,28 @@
+import { MailGeneral } from '.';
+import { getNextPaidAndValidatedUserBatch } from '../../operations/user';
+import { getNotOnDiscordServerUsers, getNotPaidSsbuUsers, getNotPaidUsers } from './targets';
+
+export const availableGeneralMails: {
+  [key: string]: MailGeneral;
+} = {
+  joindiscord: {
+    targets: getNotOnDiscordServerUsers,
+    template: 'joindiscord',
+  },
+  minor: {
+    targets: getNotPaidUsers,
+    template: 'minor',
+  },
+  notpaid: {
+    targets: getNotPaidUsers,
+    template: 'notpaid',
+  },
+  notpaidssbu: {
+    targets: getNotPaidSsbuUsers,
+    template: 'notpaidssbu',
+  },
+  tickets: {
+    targets: getNextPaidAndValidatedUserBatch,
+    template: 'tickets',
+  },
+};
